Refactor FXIOS-7216 [v119] Remove autopush feature flag
