+++ Account/SyncAuthState.swift
@@ -104,6 +104,28 @@ extension SyncAuthStateCache: JSONLiteralConvertible {
     }
 }
 
+/// Moved from an old implementation of push crypto
+/// Used by the `token` function to get base64 data from the
+/// sync keys. Will be removed in https://github.com/mozilla-mobile/firefox-ios/issues/16005
+extension String {
+    /// Returns a base64 url safe decoding of the given string.
+    /// The string is allowed to be padded
+    /// What is padding?: http://stackoverflow.com/a/26632221
+    var base64urlSafeDecodedData: Data? {
+        // We call this method twice: once with the last two args as nil, 0 – this gets us the length
+        // of the decoded string.
+        let length = ece_base64url_decode(self, self.count, ECE_BASE64URL_REJECT_PADDING, nil, 0)
+        guard length > 0 else { return nil }
+
+        // The second time, we actually decode, and copy it into a made to measure byte array.
+        var bytes = [UInt8](repeating: 0, count: length)
+        let checkLength = ece_base64url_decode(self, self.count, ECE_BASE64URL_REJECT_PADDING, &bytes, length)
+        guard checkLength == length else { return nil }
+
+        return Data(bytes: bytes, count: length)
+    }
+}
+
 open class FirefoxAccountSyncAuthState: SyncAuthState {
     private var logger: Logger
     fileprivate let cache: KeychainCache<SyncAuthStateCache>
