+++ Client/FeatureFlags/NimbusFlaggableFeature.swift
@@ -10,7 +10,6 @@ import UIKit
 /// An enum describing the featureID of all features found in Nimbus.
 /// Please add new features alphabetically.
 enum NimbusFeatureFlagID: String, CaseIterable {
-    case autopushFeature
     case bottomSearchBar
     case contextualHintForToolbar
     case creditCardAutofillStatus
@@ -55,8 +54,6 @@ struct NimbusFlaggableFeature: HasNimbusSearchBar {
         typealias FlagKeys = PrefsKeys.FeatureFlags
 
         switch featureID {
-        case .autopushFeature:
-            return FlagKeys.AutopushFeature
         case .bottomSearchBar:
             return FlagKeys.SearchBarPosition
         case .historyHighlights:
